Added Android Java implementation of Chapter 10: 'MultiTexture' sample that demonstrates using multiple textures in OpenGL ES 3.0.
This sample loads shaders (vertex and fragment) from 'assets' folder. This facility of resource loading from assets is reused for texture image loading as well.
However, this facility has not been added to Common and might be moved there in the near future as the need arises for reusing this functionality in other samples.

Currently tested with Google Nexus 4 (Qualcomm Adreno 320). Will be later tested on Nexus 7 (Gen 2) and Nexus 10.