Added "Android" folder to Chapter 14 sample: Noise3D to make it work with Android NDK.
However, it is currently NOT WORKING on LG Google Nexus 4 (Adreno 320) with the following error :

<core_glTexImage3DOES:662>: GL_INVALID_OPERATION

Probably, the Adreno driver does not properly support 3D Textures. This still needs to be investigated whether the issue is device specific or something that is reproducible on other devices as well. This will become clear once the sample will be tested on Nexus 7 and/or Nexus 10 etc.